Create Revision
This is a revision of the rough draft for the stages of writing assignment. I split up the last paragraph into two paragraphs, for a total of three body paragraphs. I also changed the last sentence of the conclusion and the last sentence of the introduction to better fit the content of the paper.